requirements.txt
asgiref==3.3.4
Django==3.2.4
gunicorn==20.1.0
Pillow==8.2.0
python-dotenv==0.18.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
psycopg2-binary==2.8.5
